修改：使用枚举代替繁琐的转换

+++ api-common-utils/src/main/java/com/kuretru/api/common/entity/data/WeappUserDO.java
@@ -20,7 +20,7 @@ public class WeappUserDO extends BaseDO {
 
     private String avatarUrl;
 
-    private Short gender;
+    private Integer gender;
 
     private String country;
